requirements.txt
black==19.10b0
GitPython==3.0.5
requests==2.22.0
tqdm==4.39.0
